pyproject.toml
[project]
name = "solomia"
version = "0.1.0"
description = "weight-loss telegram bot"
authors = [
    {name = "Anastasiia Rusakova",email = "arusakova917@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
